+++ src/main/java/dev/tonimatas/api/data/DailyInfo.java
@@ -1,4 +1,4 @@
-package dev.tonimatas.api;
+package dev.tonimatas.api.data;
 
 import dev.tonimatas.util.TimeUtils;
