+++ src/main/java/dev/tonimatas/api/data/UserSettings.java
@@ -1,4 +1,4 @@
-package dev.tonimatas.api;
+package dev.tonimatas.api.data;
 
 public class UserSettings {
     private boolean notifyDaily = false;
